# Backend/openapi.yaml
openapi: "3.1.0"
info:
  title: Kovara Indexer API
  version: "0.1.0"
  description: REST API for querying indexed on-chain data from the Kovara protocol.
servers:
  - url: http://localhost:3001/api
    description: Local development server
paths:
...
  /pools/{id}:
    get:
      summary: Get pool by ID
      operationId: getPool
      tags: [Pools]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Pool symbol identifier.
      responses:
        "200":
          description: Pool found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pool"
        "404":
          $ref: "#/components/responses/NotFound"
components:
...
  schemas:
...
    Pool:
      type: object
      required: [pool_id, token, balance, admins, threshold, created_ledger, updated_ledger]
      properties:
        pool_id:
          type: string
        token:
          type: string
        balance:
          type: string
          description: Current balance serialised as a string.
        admins:
          type: array
          items:
            type: string
        threshold:
          type: integer
          description: Minimum number of admin signatures required for withdrawal.
        created_ledger:
          type: integer
        updated_ledger:
          type: integer
...
    Error:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
        code:
          type: string
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: Resource not found
            code: NOT_FOUND
...